+++ packages/database/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.4.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.13.0
+  - @volontariapp/shared@0.9.0
+
 ## 3.4.8
 
 ### Patch Changes