+++ packages/domain-post/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.6.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.13.0
+  - @volontariapp/database@3.4.9
+  - @volontariapp/outbox@0.9.46
+
 ## 3.6.16
 
 ### Patch Changes